Guards and login mechanics
Register users are in localStorage, after login you can access your personal data and saved annoucements, but you don't have access to login and registry page.
Consider changing something in layout when user is logged in, for example show his username on toolbar, because now it's not clear if he is logged or not.